refactor(ui): extract runner catalog hook
Move runner spec, group, and policy form state and handlers out of App while keeping admin data loading in place.
